[1.0] sfSympalPlugin: Fixes to extending classes via events and askConfirmation()
git-svn-id: http://svn.symfony-project.org/plugins/sfSympalPlugin/trunk@17687 ee427ae8-e902-0410-961c-c3ed070cd9f9

+++ lib/plugins/sfSympalMenuPlugin/modules/sympal_menu_items/lib/Basesympal_menu_itemsActions.class.php
@@ -147,12 +147,21 @@ class Basesympal_menu_itemsActions extends autoSympal_menu_itemsActions
     {
       $this->redirect('@sympal_menu_manager_tree_delete?slug='.$menuItem['slug']);
     }
-    $this->askConfirmation('Are you sure?', 'Are you sure you wish to delete this menu item? This action is irreversible!');
 
-    $menuItem->getNode()->delete();
+    if ($this->askConfirmation('Are you sure?', 'Are you sure you wish to delete this menu item? This action is irreversible!'))
+    {
+      if ($menuItem->getNode()->isValidNode())
+      {
+        $menuItem->getNode()->delete();
+      }
+      else
+      {
+        $menuItem->delete();
+      }
 
-    $this->getUser()->setFlash('notice', 'Menu item was successfully deleted!');
-    $this->redirect('@sympal_menu_manager_tree?slug='.$request->getParameter('root_slug'));
+      $this->getUser()->setFlash('notice', 'Menu item was successfully deleted!');
+      $this->redirect('@sympal_menu_manager_tree?slug='.$request->getParameter('root_slug'));
+    }
   }
 
   public function executeManager_delete(sfWebRequest $request)
@@ -166,13 +175,13 @@ class Basesympal_menu_itemsActions extends autoSympal_menu_itemsActions
     
     $this->askConfirmation('Are you sure?', 'Are you sure you wish to delete this menu? This action is irreversible!');
 
-    if ($object->getNode()->isValidNode())
+    if ($menuItem->getNode()->isValidNode())
     {
-      $object->getNode()->delete();
+      $menuItem->getNode()->delete();
     }
     else
     {
-      $object->delete();
+      $menuItem->delete();
     }
 
     $this->getUser()->setFlash('notice', 'Menu was successfully deleted!');
@@ -225,6 +234,8 @@ class Basesympal_menu_itemsActions extends autoSympal_menu_itemsActions
   public function executeDelete(sfWebRequest $request)
   {
     $request->checkCSRFProtection();
+
+    $this->askConfirmation('Are you sure?', 'Are you sure you wish to delete this menu item? This action is irreversible!');
 
     $this->dispatcher->notify(new sfEvent($this, 'admin.delete_object', array('object' => $this->getRoute()->getObject())));
